requirements.txt
tensorflow>=2.12
numpy
pandas
matplotlib
scikit-learn
opencv-python
Pillow
seaborn
jupyterlab
streamlit
tensorflow-hub
tqdm
scikit-image
